+++ src/analysisd/labels.c
@@ -10,7 +10,7 @@
  */
 
 #include "headers/shared.h"
-#include "wazuh_db/wdb_global_helpers.h"
+#include "wazuh_db/helpers/wdb_global_helpers.h"
 #include "eventinfo.h"
 #include "config.h"
 #include "labels.h"
