+++ src/headers/remoted_op.h
@@ -9,7 +9,7 @@
  */
 
 #include "shared.h"
-#include "wazuh_db/wdb_global_helpers.h"
+#include "wazuh_db/helpers/wdb_global_helpers.h"
 
 
 /**
