+++ src/wazuh_db/helpers/wdb_global_helpers.c
