+++ src/wazuh_db/wdb_fim.c
@@ -10,7 +10,7 @@
  */
 
 #include "wdb.h"
-#include "wdb_global_helpers.h"
+#include "helpers/wdb_global_helpers.h"
 
 static const char *SQL_INSERT_EVENT = "INSERT INTO fim_event (id_file, type, date, size, perm, uid, gid, md5, sha1, uname, gname, mtime, inode, sha256, attributes) VALUES (?, ?, datetime(?, 'unixepoch', 'localtime'), ?, ?, ?, ?, ?, ?, ?, ?, datetime(?, 'unixepoch', 'localtime'), ?, ?, ?);";
 static const char *SQL_INSERT_FILE = "INSERT INTO fim_file (path, type) VALUES (?, ?);";
