+++ src/wazuh_modules/agent_upgrade/manager/wm_agent_upgrade_commands.c
@@ -22,7 +22,7 @@
 #include "wm_agent_upgrade_tasks.h"
 #include "wm_agent_upgrade_validate.h"
 #include "wm_agent_upgrade_upgrades.h"
-#include "wazuh_db/wdb_global_helpers.h"
+#include "wazuh_db/helpers/wdb_global_helpers.h"
 
 /**
  * Analyze agent information and returns a JSON to be sent to the task manager
