+++ src/wazuh_modules/vulnerability_detector/wm_vuln_detector.c
@@ -15,7 +15,7 @@
 #include "wm_vuln_detector_db.h"
 #include "wm_vuln_detector_evr.h"
 #include "addagent/manage_agents.h"
-#include "wazuh_db/wdb_global_helpers.h"
+#include "wazuh_db/helpers/wdb_global_helpers.h"
 #include <netinet/tcp.h>
 #include <openssl/ssl.h>
 #include <os_net/os_net.h>
