+++ src/wazuh_modules/vulnerability_detector/wm_vuln_detector_nvd.c
@@ -4,7 +4,7 @@
 #include "wm_vuln_detector_db.h"
 #include "wm_vuln_detector_evr.h"
 #include <os_net/os_net.h>
-#include "wazuh_db/wdb_global_helpers.h"
+#include "wazuh_db/helpers/wdb_global_helpers.h"
 
 #ifdef WAZUH_UNIT_TESTING
 // Remove STATIC qualifier from tests
